requirements.txt
mcp
aiohttp
beautifulsoup4
python-dotenv
nest-asyncio
torch
transformers
accelerate